Implemented.
Uses 3 Gains; one of them serves as the hardwired "inner plugin", in this version just muting one of the signals. The other 2 gains implement the mixing knob behavior.
All the processors are connected using the AudioProcessorGraph. That however doesn't produce the expected result, the ouput plays very badly unless one of the signals being combined is completely muted.